change Double to BigDecimal in Transaction and finance tools

+++ app/src/main/java/com/rygital/moneytracker/ui/dashboard/Dashboard.kt
@@ -2,11 +2,12 @@ package com.rygital.moneytracker.ui.dashboard
 
 import com.rygital.moneytracker.ui.base.MvpPresenter
 import com.rygital.moneytracker.ui.base.MvpView
+import java.math.BigDecimal
 
 interface Dashboard {
     interface View: MvpView {
-        fun showMoneyInRUB(value: Double)
-        fun showMoneyInUSD(value: Double)
+        fun showMoneyInRUB(value: BigDecimal)
+        fun showMoneyInUSD(value: BigDecimal)
     }
 
     interface Presenter<in V: View>: MvpPresenter<V> {
